+++ bin/ensembl2uniprot.py
@@ -7,7 +7,7 @@
 
 import argparse
 
-from utils import check_id_mapping_results_ready, safe_get, safe_post
+from utils import check_id_mapping_results_ready, handle_http_request
 
 
 def ensembl2uniprot(ensembl_ids: list[str]) -> list[str]:
@@ -15,25 +15,16 @@ def ensembl2uniprot(ensembl_ids: list[str]) -> list[str]:
     if len(ensembl_ids) == 0:
         return []
 
-    payload = {
-        "ids": ensembl_ids,
-        "from": "Ensembl",
-        "to": "UniProtKB"
-    }
+    payload = {"ids": ensembl_ids, "from": "Ensembl", "to": "UniProtKB"}
 
-    res = safe_post("https://rest.uniprot.org/idmapping/run", data=payload)
+    json = handle_http_request("https://rest.uniprot.org/idmapping/run", method="POST", data=payload)
 
-    if not res.ok:
-        raise ValueError(f"HTTP error: {res.status_code}")
-
-    job_id = res.json()["jobId"]
+    job_id = json["jobId"]
 
     # wait for the job to finish
     check_id_mapping_results_ready(job_id)
 
-    res = safe_get(f"https://rest.uniprot.org/idmapping/results/{job_id}")
-
-    json = res.json()
+    json = handle_http_request(f"https://rest.uniprot.org/idmapping/results/{job_id}")
 
     mapped_ids = [i["from"] for i in json["results"] if len(i["to"]) > 0]
     unmapped_ids = [i for i in ensembl_ids if i not in mapped_ids]
@@ -43,9 +34,7 @@ def ensembl2uniprot(ensembl_ids: list[str]) -> list[str]:
 
 
 def main() -> None:
-    parser = argparse.ArgumentParser(
-        description="Convert Ensembl IDs to UniProt IDs using the UniProt mapping API."
-    )
+    parser = argparse.ArgumentParser(description="Convert Ensembl IDs to UniProt IDs using the UniProt mapping API.")
     parser.add_argument(
         "-i",
         "--id",
@@ -55,6 +44,7 @@ def main() -> None:
     args = parser.parse_args()
 
     print(ensembl2uniprot([args.id]))
+
 
 if __name__ == "__main__":
     main()