+++ modules/local/fetch_ensembl_idmap/resources/usr/bin/fetch_ensembl_idmap.py
@@ -10,7 +10,7 @@ import requests
 
 def main() -> None:
     headers = {"content-type": "application/json"}
-    res = requests.get("https://rest.ensembl.org/info/species", headers = headers)
+    res = requests.get("https://rest.ensembl.org/info/species", headers=headers)
 
     for entry in res.json()["species"]:
         print(f"{entry['name']},{entry['taxon_id']}")