+++ modules/local/fetch_oma_group_online/resources/usr/bin/fetch_oma_groupid.py
@@ -5,23 +5,19 @@
 
 """Get OMA group ID from a UniProt ID."""
 
-import os
 import argparse
+import os
 import subprocess
 import sys
-from warnings import warn
 
-bin_dir = os.path.dirname(os.path.realpath(subprocess.check_output(
-    ['which', 'utils.py'], text=True).strip()))
+bin_dir = os.path.dirname(os.path.realpath(subprocess.check_output(["which", "utils.py"], text=True).strip()))
 sys.path.insert(0, bin_dir)
 
-from utils import safe_get # noqa: E402
+from utils import handle_http_request  # noqa: E402
 
 
 def main() -> None:
-    parser = argparse.ArgumentParser(
-        description="Get OMA group ID from a UniProt ID."
-    )
+    parser = argparse.ArgumentParser(description="Get OMA group ID from a UniProt ID.")
     parser.add_argument(
         "-p",
         "--protein-id",
@@ -32,16 +28,12 @@ def main() -> None:
 
     prot_id = args.protein_id
     headers = {"User-Agent": "pyomadb/2.1.0"}
-    res = safe_get(f"https://omabrowser.org/api/protein/{prot_id}", headers=headers)
+    json = handle_http_request(f"https://omabrowser.org/api/protein/{prot_id}", headers=headers)
 
-    if res.status_code == 404:
-        warn("ID not found")
+    if json == dict() or "oma_group" not in json:
         print("0")
         return
-    elif not res.ok:
-        raise ValueError("Fetch failed, aborting")
 
-    json = res.json()
     entry: dict = dict()
     if json["is_main_isoform"]:
         entry = json
@@ -49,15 +41,14 @@ def main() -> None:
     # If main isoform not found, check the first alternative isoform
     if entry == dict():
         if len(json["alternative_isoforms_urls"]) > 0:
-            res = safe_get(json["isoforms"], headers=headers)
-            json2 = res.json()
+            json2 = handle_http_request(json["isoforms"], headers=headers)
             for isoform in json2:
                 if isoform["is_main_isoform"]:
                     entry = isoform
                     break
             if entry == dict():
                 raise ValueError("Isoform not found")
-    print(entry['oma_group'])
+    print(entry["oma_group"])
 
 
 if __name__ == "__main__":