+++ modules/local/fetch_panther_group_online/resources/usr/bin/fetch_panther_group.py
@@ -5,23 +5,20 @@
 
 """Fetch members of a Panther group by ID."""
 
-import os
 import argparse
+import os
 import subprocess
 import sys
 from warnings import warn
 
-bin_dir = os.path.dirname(os.path.realpath(subprocess.check_output(
-    ['which', 'utils.py'], text=True).strip()))
+bin_dir = os.path.dirname(os.path.realpath(subprocess.check_output(["which", "utils.py"], text=True).strip()))
 sys.path.insert(0, bin_dir)
 
-from utils import safe_get # noqa: E402
+from utils import handle_http_request  # noqa: E402
 
 
 def main() -> None:
-    parser = argparse.ArgumentParser(
-        description="Fetch members of a Panther group by ID."
-    )
+    parser = argparse.ArgumentParser(description="Fetch members of a Panther group by ID.")
     parser.add_argument(
         "-i",
         "--input-id",
@@ -36,24 +33,24 @@ def main() -> None:
     )
     args = parser.parse_args()
 
-    res = safe_get(f"https://www.pantherdb.org/services/oai/pantherdb/ortholog/matchortho?geneInputList={args.input_id}&organism={args.organism}&orthologType=all")
+    url = f"https://www.pantherdb.org/services/oai/pantherdb/ortholog/matchortho?geneInputList={args.input_id}&organism={args.organism}&orthologType=all"
+    json = handle_http_request(url)
 
-    if not res.ok:
-        raise ValueError(f"HTTP error: {res.status_code}")
-
-    json = res.json()
     try:
         for i in json["search"]["mapping"]["mapped"]:
             uniprot_id = i["target_gene"].split("|")[-1].split("=")[-1]
             print(f"{uniprot_id}")
     except KeyError:
         warn("No results found")
-        pass # yes, I mean this, we just want to return an empty file if nothing is found
+        pass  # yes, I mean this, we just want to return an empty file if nothing is found
 
     try:
-        print(f"{json['search']['product']['content']} {json['search']['product']['version']}", file="panther_version.txt")
+        print(
+            f"{json['search']['product']['content']} {json['search']['product']['version']}", file="panther_version.txt"
+        )
     except KeyError:
         print("error", file="panther_version.txt")
+
 
 if __name__ == "__main__":
     main()