+++ modules/local/filter_hits/resources/usr/bin/score_hits.py
@@ -3,9 +3,8 @@
 # Written by Igor Trujnara, released under the MIT license
 # See https://opensource.org/license/mit for details
 
-import csv
 import argparse
-import sys
+import csv
 
 
 def load_data_from_csv(file_path) -> list:
@@ -24,7 +23,7 @@ def filter_data(data, threshold) -> list:
     """
     filtered_data = []
     for row in data:
-        if float(row['score']) >= threshold:
+        if float(row["score"]) >= threshold:
             filtered_data.append(row)
     return filtered_data
 
@@ -34,7 +33,7 @@ def filter_centroid(data) -> list:
     Find the centroid (highest agreement) source and filter data by it.
     """
     # get columns except first two and last one into a list of lists
-    columns = [[float(list(row.values())[i]) for row in data] for i in range(2, len(data[0])-1)]
+    columns = [[float(list(row.values())[i]) for row in data] for i in range(2, len(data[0]) - 1)]
 
     # calculate agreement
     scores = [0 for _ in columns]
@@ -50,7 +49,7 @@ def filter_centroid(data) -> list:
     # filter data
     filtered_data = []
     for i in range(len(data)):
-        if list(data[i].values())[centroid+1] == '1':
+        if list(data[i].values())[centroid + 1] == "1":
             filtered_data.append(data[i])
     return filtered_data
 
@@ -88,22 +87,22 @@ def main():
         query = f.read().strip()
 
     # filter data
-    for score in range(1, max([int(row['score']) for row in data])+1):
-        f = open(f"{prefix}_minscore_{score}.txt", 'w')
+    for score in range(1, max([int(row["score"]) for row in data]) + 1):
+        f = open(f"{prefix}_minscore_{score}.txt", "w")
         filtered_data = filter_data(data, score)
         print(query, file=f)
         for row in filtered_data:
-            print(row['id'], file=f)
+            print(row["id"], file=f)
         f.close()
 
     filtered_data = filter_centroid(data)
 
-    f = open(f"{prefix}_centroid.txt", 'w')
+    f = open(f"{prefix}_centroid.txt", "w")
 
     print(query, file=f)
 
     for row in filtered_data:
-        print(row['id'], file=f)
+        print(row["id"], file=f)
     f.close()