+++ modules/local/write_seqinfo/resources/usr/bin/fetch_oma_taxid_by_id.py
@@ -5,17 +5,16 @@
 
 """Fetch OMA taxon ID by UniProt ID."""
 
-import os
 import argparse
+import os
 import subprocess
 import sys
 from warnings import warn
 
-bin_dir = os.path.dirname(os.path.realpath(subprocess.check_output(
-    ['which', 'utils.py'], text=True).strip()))
+bin_dir = os.path.dirname(os.path.realpath(subprocess.check_output(["which", "utils.py"], text=True).strip()))
 sys.path.insert(0, bin_dir)
 
-from utils import safe_get # noqa: E402
+from utils import handle_http_request  # noqa: E402
 
 
 def main() -> None:
@@ -29,19 +28,18 @@ def main() -> None:
     args = parser.parse_args()
 
     uniprot_id = args.uniprot_id
-    headers = {"User-Agent": "pyomadb/2.1.0" }
-    res = safe_get(f"https://omabrowser.org/api/protein/{uniprot_id}/", headers=headers)
+    headers = {"User-Agent": "pyomadb/2.1.0"}
+    url = f"https://omabrowser.org/api/protein/{uniprot_id}/"
+    json = handle_http_request(url, headers=headers)
 
-    if res.status_code == 404:
+    if not json:
         warn("ID not found")
         print("1")
-    elif not res.ok:
-        raise ValueError(f"Fetch failed (HTTP {res.status_code}), aborting\nResponse content: {res.headers}\n{res.text}")
 
     try:
-        print(res.json()["species"]["taxon_id"])
+        print(json["species"]["taxon_id"])
     except KeyError:
-        print("1") # default to root if no taxid is found
+        print("1")  # default to root if no taxid is found
 
 
 if __name__ == "__main__":